Move artisticFilter to simpleEffect
